move fetchIngredients to its own script
