+++ frontend/www/ingredients.html
@@ -72,7 +72,7 @@
 
     <script src="scripts/defaults.js"></script>
     <script src="scripts/scripts.js"></script>
-    <script src="scripts/meals/ingredients-page.js"></script>
+    <script type="module" src="scripts/meals/ingredients-page.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
